e2e/pyproject.toml
[project]
name = "kubetail-e2e"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "grpcio>=1.60",
    "pytest>=8.0",
    "python-dotenv>=1.0",
    "requests>=2.31",
    "websockets>=13.0",
]

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::urllib3.exceptions.InsecureRequestWarning",
]
